requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.3.1
idna==3.3
iniconfig==1.1.1
mysqlclient==2.1.1
packaging==21.3
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.7
virtualenv==20.9.0
